+++ cardano-node-socket-emulator/src/Cardano/Node/Socket/Emulator/Server.hs
@@ -1,89 +1,482 @@
-{-# LANGUAGE FlexibleContexts   #-}
-{-# LANGUAGE NamedFieldPuns     #-}
-{-# LANGUAGE NumericUnderscores #-}
-{-# LANGUAGE OverloadedStrings  #-}
-{-# LANGUAGE TypeApplications   #-}
+{-# LANGUAGE DataKinds         #-}
+{-# LANGUAGE FlexibleContexts  #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GADTs             #-}
+{-# LANGUAGE LambdaCase        #-}
+{-# LANGUAGE NamedFieldPuns    #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeFamilies      #-}
 
-module Cardano.Node.Socket.Emulator.Server
-    ( main
-    ) where
+module Cardano.Node.Socket.Emulator.Server (ServerHandler, runServerNode, processBlock, modifySlot, addTx) where
 
 import Cardano.BM.Data.Trace (Trace)
-import Cardano.Node.Emulator.Internal.Node (Params (..), SlotConfig (SlotConfig, scSlotLength, scSlotZeroTime))
-import Cardano.Node.Socket.Emulator qualified as Server
-import Cardano.Node.Socket.Emulator.API (API)
-import Cardano.Node.Socket.Emulator.Mock (consumeEventHistory, healthcheck, processChainEffects, slotCoordinator)
-import Cardano.Node.Socket.Emulator.Params qualified as Params
-import Cardano.Node.Socket.Emulator.Types (AppState (..), NodeServerConfig (..), PABServerLogMsg (..),
-                                           initialChainState)
-import Control.Concurrent (MVar, forkIO, newMVar)
-import Control.Monad (void)
-import Control.Monad.Freer.Extras.Delay (delayThread, handleDelayEffect)
-import Control.Monad.Freer.Extras.Log (logInfo)
-import Control.Monad.IO.Class (liftIO)
-import Data.Function ((&))
-import Data.Map.Strict qualified as Map
-import Data.Proxy (Proxy (Proxy))
-import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
-import Data.Time.Units (Millisecond, Second)
-import Ledger.Value.CardanoAPI qualified as CardanoAPI
-import Network.Wai.Handler.Warp qualified as Warp
+import Data.ByteString.Lazy qualified as LBS
+import Data.List (intersect)
+import Data.Maybe (listToMaybe)
+import Data.Void (Void)
+
+import Control.Concurrent
+import Control.Concurrent.Async
+import Control.Concurrent.STM
+import Control.Lens hiding (index, ix)
+import Control.Monad.Freer (interpret, runM)
+import Control.Monad.Freer.State (runState)
+import Control.Monad.Reader
+import Control.Tracer
+
+import Ouroboros.Network.Protocol.ChainSync.Server (ChainSyncServer (..), ServerStIdle (..), ServerStIntersect (..),
+                                                    ServerStNext (..))
+import Ouroboros.Network.Protocol.ChainSync.Server qualified as ChainSync
+import Ouroboros.Network.Protocol.LocalTxSubmission.Server qualified as TxSubmission
+import Ouroboros.Network.Protocol.LocalTxSubmission.Type qualified as TxSubmission
 import Plutus.Monitoring.Util qualified as LM
-import Servant (Application, hoistServer, serve, (:<|>) ((:<|>)))
-import Servant.Client (BaseUrl (baseUrlPort))
-import Wallet.Emulator.Wallet (fromWalletNumber)
 
-app ::
-    Trace IO PABServerLogMsg
+import Cardano.Slotting.Slot (SlotNo (..), WithOrigin (..))
+import Ouroboros.Network.Block (Point (..), pointSlot)
+import Ouroboros.Network.IOManager
+import Ouroboros.Network.Mux
+import Ouroboros.Network.NodeToClient (NodeToClientProtocols (..), nodeToClientCodecCBORTerm,
+                                       nodeToClientHandshakeCodec, nullErrorPolicies, versionedNodeToClientProtocols)
+import Ouroboros.Network.Point qualified as OP (Block (..))
+import Ouroboros.Network.Protocol.Handshake.Codec
+import Ouroboros.Network.Protocol.Handshake.Version
+import Ouroboros.Network.Snocket
+import Ouroboros.Network.Socket
+
+import Cardano.Api qualified as C
+
+import Cardano.Protocol.Socket.Type
+
+import Cardano.Node.Emulator.Internal.Node (ChainEvent, Params)
+import Cardano.Node.Emulator.Internal.Node.Chain qualified as Chain
+import Cardano.Node.Socket.Emulator.Chain (MockNodeServerChainState (..), addTxToPool, chainNewestFirst, channel,
+                                           currentSlot, getChannel, getTip, handleControlChain, tip, txPool)
+import Ledger (Block, CardanoTx (..), Slot (..))
+
+data CommandChannel = CommandChannel
+  { ccCommand  :: TQueue ServerCommand
+  , ccResponse :: TQueue ServerResponse
+  }
+
+{- | Clone the original channel for each connected client, then use
+     this wrapper to make sure that no data is consumed from the
+     original channel. -}
+newtype LocalChannel = LocalChannel (TChan Block)
+
+{- | A handler used to pass around the path to the server
+     and channels used for controlling the server. -}
+data ServerHandler = ServerHandler {
+    shSocketPath     :: FilePath,
+    -- The client will send a `ServerCommand` and the server will
+    -- respond with a `ServerResponse`.
+    shCommandChannel :: CommandChannel
+}
+
+{- | The commands that control the server. This API is not part of the client
+     interface, and in order to call them directly you will need access to the
+     returned ServerHandler -}
+data ServerCommand =
+    -- This command will add a new block by processing
+    -- transactions in the memory pool.
+    ProcessBlock
+    -- Set the slot number
+  | ModifySlot (Slot -> Slot)
+    -- Append a transaction to the transaction pool.
+  | AddTx (C.Tx C.BabbageEra)
+
+instance Show ServerCommand where
+    show = \case
+        ProcessBlock -> "ProcessBlock"
+        ModifySlot _ -> "ModifySlot"
+        AddTx t      -> "AddTx " <> show t
+
+{- | The response from the server. Can be used for the information
+     passed back, or for synchronisation.
+-}
+data ServerResponse =
+    -- A block was added. We are using this for synchronization.
+    BlockAdded Block
+    | SlotChanged Slot
+    deriving Show
+
+processBlock :: MonadIO m => ServerHandler -> m Block
+processBlock ServerHandler {shCommandChannel} = do
+    liftIO $ atomically $ writeTQueue (ccCommand shCommandChannel) ProcessBlock
+    -- Wait for the server to finish processing blocks.
+    liftIO $ atomically $ readTQueue (ccResponse shCommandChannel) >>= \case
+        BlockAdded block -> pure block
+        _                -> retry
+
+modifySlot :: MonadIO m => (Slot -> Slot) -> ServerHandler -> m Slot
+modifySlot f ServerHandler{shCommandChannel} = do
+    liftIO $ atomically $ writeTQueue (ccCommand shCommandChannel) $ ModifySlot f
+    -- Wait for the server to finish changing the slot.
+    liftIO $ atomically $ readTQueue (ccResponse shCommandChannel) >>= \case
+        SlotChanged slot -> pure slot
+        _                -> retry
+
+addTx :: MonadIO m => ServerHandler -> C.Tx C.BabbageEra -> m ()
+addTx ServerHandler { shCommandChannel } tx = do
+    liftIO $ atomically $ writeTQueue (ccCommand  shCommandChannel) $ AddTx tx
+
+{- Create a thread that keeps the number of blocks in the channel to the maximum
+   limit of K -}
+pruneChain :: MonadIO m => Integer -> TChan Block -> m ThreadId
+pruneChain k original = do
+  localChannel <- liftIO $ atomically $ cloneTChan original
+  liftIO . forkIO $ go k localChannel
+  where
+  go :: MonadIO m => Integer -> TChan Block -> m ()
+  go k' localChannel = do
+    -- Wait for data on the channel
+    _ <- liftIO $ atomically $ readTChan localChannel
+    if k' == 0
+       {- When the counter reaches zero, there are K blocks in the
+          original channel and we start to remove the oldest stored
+          block by reading it. -}
+       then do
+           liftIO $ atomically (readTChan original) >> go 0 localChannel
+       else do
+           go (k' - 1) localChannel
+
+handleCommand ::
+    MonadIO m
+ => Trace IO ChainEvent
+ -> CommandChannel
+ -> MVar MockNodeServerChainState
  -> Params
- -> MVar AppState
- -> Application
-app trace params stateVar =
-    serve (Proxy @API) $
-    hoistServer
-        (Proxy @API)
-        (liftIO . processChainEffects (LM.convertLog ProcessingChainEvent trace) params stateVar)
-        (healthcheck :<|> (fmap (fmap ProcessingChainEvent) <$> consumeEventHistory stateVar))
+ -> m ()
+handleCommand trace CommandChannel {ccCommand, ccResponse} mvChainState params =
+    liftIO (atomically $ readTQueue ccCommand) >>= \case
+        AddTx tx     -> do
+            liftIO $ modifyMVar_ mvChainState (pure . over txPool (CardanoEmulatorEraTx tx :))
+        ModifySlot f -> liftIO $ do
+            state <- liftIO $ takeMVar mvChainState
+            (s, nextState') <- liftIO $ Chain.modifySlot f
+                  & interpret (handleControlChain params)
+                  & interpret (LM.handleLogMsgTrace trace)
+                  & runState state
+                  & runM
+            putMVar mvChainState nextState'
+            atomically $ do
+                writeTQueue ccResponse (SlotChanged s)
+        ProcessBlock -> liftIO $ do
+            state <- liftIO $ takeMVar mvChainState
+            (block, nextState') <- liftIO $ Chain.processBlock
+                  & interpret (handleControlChain params)
+                  & interpret (LM.handleLogMsgTrace trace)
+                  & runState state
+                  & runM
+            putMVar mvChainState nextState'
+            atomically $
+                writeTQueue ccResponse (BlockAdded block)
 
-data Ctx = Ctx { serverHandler :: Server.ServerHandler
-               , serverState   :: MVar AppState
-               , mockTrace     :: Trace IO PABServerLogMsg
-               }
+{- | Start the server in a new thread, and return a server handler
+     used to control the server -}
+runServerNode ::
+    MonadIO m
+ => Trace IO ChainEvent
+ -> FilePath
+ -> Integer
+ -> MockNodeServerChainState
+ -> Params
+ -> m ServerHandler
+runServerNode trace shSocketPath k initialState params = liftIO $ do
+    serverState      <- newMVar initialState
+    shCommandChannel <- CommandChannel <$> newTQueueIO <*> newTQueueIO
+    globalChannel    <- getChannel serverState
+    void $ forkIO . void    $ protocolLoop        shSocketPath     serverState
+    void $ forkIO . forever $ handleCommand trace shCommandChannel serverState params
+    void                    $ pruneChain k globalChannel
+    pure $ ServerHandler { shSocketPath, shCommandChannel }
 
-main :: Trace IO PABServerLogMsg -> NodeServerConfig -> IO () -> IO ()
-main trace nodeServerConfig@NodeServerConfig { nscBaseUrl
-                            , nscSlotConfig
-                            , nscKeptBlocks
-                            , nscInitialTxWallets
-                            , nscSocketPath } whenStarted = LM.runLogEffects trace $ do
+-- * ChainSync protocol
 
-    -- make initial distribution of 1 billion Ada to all configured wallets
-    let dist = Map.fromList $ zip (fromWalletNumber <$> nscInitialTxWallets) (repeat (CardanoAPI.adaValueOf 1_000_000_000))
-    initialState <- initialChainState dist
-    let appState = AppState
-            { _chainState = initialState
-            , _eventHistory = mempty
-            }
-    params <- liftIO $ Params.fromNodeServerConfig nodeServerConfig
-    serverHandler <- liftIO $ Server.runServerNode (LM.convertLog ProcessingChainEvent trace) nscSocketPath nscKeptBlocks (_chainState appState) params
-    serverState   <- liftIO $ newMVar appState
-    handleDelayEffect $ delayThread (2 :: Second)
+{- A monad for running all code executed when a state
+   transition is invoked. It makes the implementation of
+   state transitions easier to read. -}
 
-    let ctx = Ctx { serverHandler = serverHandler
-                  , serverState   = serverState
-                  , mockTrace     = trace
-                  }
+type ChainSyncMonad = ReaderT (MVar MockNodeServerChainState) IO
 
-    runSlotCoordinator ctx
+runChainSync :: MVar MockNodeServerChainState -> ChainSyncMonad a -> IO a
+runChainSync = flip runReaderT
 
-    logInfo $ StartingPABServer $ baseUrlPort nscBaseUrl
-    liftIO $ Warp.runSettings warpSettings $ app trace params serverState
+{- The initial state of the protocol. You can move into
+   requesting the next block or reset state by searching for an
+   intersection. -}
+idleState ::
+    ( MonadReader (MVar MockNodeServerChainState) m
+    , MonadIO m )
+ => LocalChannel
+ -> m (ServerStIdle Block (Point Block) Tip m ())
+idleState channel' =
+    pure ServerStIdle {
+        recvMsgRequestNext = nextState channel',
+        recvMsgFindIntersect = findIntersect,
+        recvMsgDoneClient = return ()
+    }
 
-        where
-            warpSettings = Warp.defaultSettings & Warp.setPort (baseUrlPort nscBaseUrl) & Warp.setBeforeMainLoop whenStarted
+{- Get the next block, either immediately (the Just/Left branch)
+   or within a monad (IO, in our case) where you can wait for the
+   next block (Nothing/Right branch) -}
+nextState ::
+    ( MonadReader (MVar MockNodeServerChainState) m
+    , MonadIO m )
+ => LocalChannel
+ -> m (Either (ServerStNext Block (Point Block) Tip m ())
+              (m (ServerStNext Block (Point Block) Tip m ())))
+nextState localChannel@(LocalChannel channel') = do
+    chainState <- ask
+    tip' <- getTip chainState
+    (liftIO . atomically $ tryReadTChan channel') >>= \case
+        Nothing -> do
+            Right . pure <$> do
+                nextBlock <- liftIO . atomically $ readTChan channel'
+                liftIO $ modifyMVar_ chainState (pure . (tip ?~ nextBlock))
+                sendRollForward localChannel tip' nextBlock
+        Just nextBlock -> do
+            liftIO $ modifyMVar_ chainState (pure . (tip ?~ nextBlock))
+            Left <$> sendRollForward localChannel tip' nextBlock
 
-            runSlotCoordinator (Ctx serverHandler _ _)  = do
-                let SlotConfig{scSlotZeroTime, scSlotLength} = nscSlotConfig
-                logInfo $ StartingSlotCoordination (posixSecondsToUTCTime $ realToFrac scSlotZeroTime / 1000)
-                                                   (fromInteger scSlotLength :: Millisecond)
-                void $ liftIO $ forkIO $ slotCoordinator nscSlotConfig serverHandler
+{- This protocol state will search for a block intersection
+   with some client provided blocks. When an intersection is found
+   the client state is reset to the new offset (the Just branch)
+   or to the genesis block if no intersection was found. -}
+findIntersect ::
+    ( MonadReader (MVar MockNodeServerChainState) m
+    , MonadIO m )
+ => [Point Block]
+ -> m (ServerStIntersect Block (Point Block) Tip m ())
+findIntersect clientPoints = do
+    mvState <- ask
+    chainState <- liftIO $ readMVar mvState
+    serverPoints <- getChainPoints (view channel chainState) chainState
+    let point = listToMaybe
+              $ intersect serverPoints
+                          clientPoints
+    tip' <- getTip mvState
+    pure $ case point of
+        Nothing ->
+          SendMsgIntersectNotFound
+            tip'
+            -- No intersection found. Resume from origin.
+            (ChainSyncServer $ cloneChainFrom 0 >>= idleState)
+        Just point' ->
+          SendMsgIntersectFound
+            point'
+            tip'
+            -- Resuming from point'.
+            (ChainSyncServer $ cloneChainFrom (pointOffset point') >>= idleState)
+
+{- This is a wrapper around the creation of a `ServerStNext` -}
+sendRollForward ::
+    ( MonadReader (MVar MockNodeServerChainState) m
+    , MonadIO m )
+ => LocalChannel
+ -> Block -- tip
+ -> Block -- current
+ -> m (ServerStNext Block (Point Block) Tip m ())
+sendRollForward channel' tip' current = pure $
+    SendMsgRollForward
+        current
+        tip'
+        (ChainSyncServer (idleState channel'))
+
+{- This is the state for a new connection. For now we start with
+   slot 0, and in idleState. This will probably change, since it
+   makes more sense to start in the `StIntersect` state. -}
+chainSyncServer ::
+    ( MonadReader (MVar MockNodeServerChainState) m
+    , MonadIO m )
+ => ChainSyncServer Block (Point Block) Tip m ()
+chainSyncServer =
+    ChainSyncServer (cloneChainFrom 0 >>= idleState)
+
+{- Use a `TChan` to model a broadcast channel of which we
+   clone (with potentially varying offsets) for clients. -}
+cloneChainFrom :: forall m.
+    ( MonadReader (MVar MockNodeServerChainState) m
+    , MonadIO m )
+ => Integer
+ -> m LocalChannel
+cloneChainFrom offset = LocalChannel <$> go
+  where
+    go :: m (TChan Block)
+    go = do
+        globalChannel <- ask >>= getChannel
+        liftIO $ atomically $ do
+            localChannel <- cloneTChan globalChannel
+            consume localChannel offset
+
+    consume :: TChan a -> Integer -> STM (TChan a)
+    consume channel' ix | ix == 0    = pure channel'
+    consume channel' ix =
+        -- We should have all requested blocks available on the
+        -- channel, for consumption.
+        tryReadTChan channel' >> consume channel' (ix - 1)
+
+-- * Protocol setup
+
+{- The node protocols always run in the IO monad. I wanted to use a
+   different monad stack (mainly to be able to pass the internal state
+   in a `MonadReader` and future proofing) so I wrote some hoisting
+   functions for each of the states which transform the `ChainSyncMonad`
+   into IO. -}
+
+hoistChainSync ::
+    MonadReader (MVar MockNodeServerChainState) m
+ => ChainSyncServer Block (Point Block) Tip ChainSyncMonad a
+ -> m (ChainSyncServer Block (Point Block) Tip IO a)
+hoistChainSync machine = do
+    internalState <- ask
+    pure ChainSyncServer {
+        {- The basic idea is running the reader monad to remove it,
+           leaving only IO, which is what we need. We do the same for all
+           other states. -}
+        runChainSyncServer = runChainSync internalState $
+            runChainSyncServer machine >>= hoistStIdle
+    }
+
+hoistStIdle ::
+    MonadReader (MVar MockNodeServerChainState) m
+ => ServerStIdle Block (Point Block) Tip ChainSyncMonad a
+ -> m (ServerStIdle Block (Point Block) Tip IO a)
+hoistStIdle (ServerStIdle nextState' findIntersect' done) = do
+    internalState <- ask
+    pure ServerStIdle {
+        recvMsgRequestNext =
+            runChainSync internalState $
+                nextState' >>= \case
+                    Left stNext -> Left         <$>  hoistStNext     stNext
+                    Right mNext -> Right . pure <$> (hoistStNext =<< mNext ),
+        recvMsgFindIntersect = \points ->
+            runChainSync internalState
+                         (findIntersect' points >>= hoistStIntersect),
+        recvMsgDoneClient    = runChainSync internalState done
+   }
+
+hoistStIntersect ::
+    MonadReader (MVar MockNodeServerChainState) m
+ => ServerStIntersect Block (Point Block) Tip ChainSyncMonad a
+ -> m (ServerStIntersect Block (Point Block) Tip IO a)
+hoistStIntersect (SendMsgIntersectFound point tip' nextState') =
+    SendMsgIntersectFound point tip' <$> hoistChainSync nextState'
+hoistStIntersect (SendMsgIntersectNotFound tip' nextState') =
+    SendMsgIntersectNotFound tip'    <$> hoistChainSync nextState'
+
+hoistStNext ::
+    MonadReader (MVar MockNodeServerChainState) m
+ => ServerStNext Block (Point Block) Tip ChainSyncMonad a
+ -> m (ServerStNext Block (Point Block) Tip IO a)
+hoistStNext (SendMsgRollForward header tip' nextState') =
+    SendMsgRollForward header tip' <$> hoistChainSync nextState'
+hoistStNext (SendMsgRollBackward header tip' nextState') =
+    SendMsgRollBackward header tip' <$> hoistChainSync nextState'
+
+{- This is boilerplate code that sets up the node protocols,
+   you can find in:
+     ouroboros-network/ouroboros-network/demo/chain-sync.hs -}
+
+protocolLoop ::
+    MonadIO m
+ => FilePath
+ -> MVar MockNodeServerChainState
+ -> m Void
+protocolLoop socketPath internalState = liftIO $ withIOManager $ \iocp -> do
+    networkState <- newNetworkMutableState
+    _ <- async $ cleanNetworkMutableState networkState
+    withServerNode
+      (localSnocket iocp)
+      nullNetworkServerTracers
+      networkState
+      (AcceptedConnectionsLimit maxBound maxBound 0)
+      (localAddressFromPath socketPath)
+      nodeToClientHandshakeCodec
+      noTimeLimitsHandshake
+      (cborTermVersionDataCodec nodeToClientCodecCBORTerm)
+      acceptableVersion
+      (SomeResponderApplication <$>
+        versionedNodeToClientProtocols
+          nodeToClientVersion
+          nodeToClientVersionData
+          (\_ _ -> nodeToClientProtocols internalState))
+      nullErrorPolicies
+      $ \_ serverAsync -> wait serverAsync
+
+nodeToClientProtocols
+  :: MVar MockNodeServerChainState
+  -> NodeToClientProtocols 'ResponderMode LBS.ByteString IO Void ()
+nodeToClientProtocols internalState =
+  NodeToClientProtocols
+    { localChainSyncProtocol = chainSync internalState
+    , localTxSubmissionProtocol = txSubmission internalState
+    , localStateQueryProtocol = doNothingResponderProtocol
+    , localTxMonitorProtocol = doNothingResponderProtocol
+    }
+
+chainSync
+  :: MVar MockNodeServerChainState
+  -> RunMiniProtocol 'ResponderMode LBS.ByteString IO Void ()
+chainSync mvChainState =
+     ResponderProtocolOnly $
+     MuxPeer
+       nullTracer
+       chainSyncCodec
+       (ChainSync.chainSyncServerPeer
+           (runReader (hoistChainSync chainSyncServer)
+                      mvChainState))
+
+txSubmission
+  :: MVar MockNodeServerChainState
+  -> RunMiniProtocol 'ResponderMode LBS.ByteString IO Void ()
+txSubmission mvChainState =
+    ResponderProtocolOnly $
+    MuxPeer
+      nullTracer
+      txSubmissionCodec
+      (TxSubmission.localTxSubmissionServerPeer
+          (pure $ txSubmissionServer mvChainState))
+
+-- * Computing intersections
+
+-- Given a `Point` find its offset into the chain.
+pointOffset :: Point Block
+            -> Integer
+pointOffset pt =
+  case pointSlot pt of
+    Origin        -> 0
+    At (SlotNo s) -> fromIntegral s
+
+-- Currently selects all points from the blockchain.
+getChainPoints :: MonadIO m => TChan Block -> MockNodeServerChainState -> m [Point Block]
+getChainPoints ch st = do
+  chain <- chainNewestFirst ch
+  pure $ zipWith mkPoint
+    [st ^. currentSlot, st ^. currentSlot - 1 .. 0]
+    chain
+  where
+    mkPoint :: Slot -> Block -> Point Block
+    mkPoint (Slot s) block =
+      Point (At (OP.Block (SlotNo $ fromIntegral s)
+                          (blockId block)))
+
+-- * TxSubmission protocol
+
+{- I did not use the same approach for this protocol as I did
+   for the `ChainSync`. This protocol has only one state and
+   it is much simpler. -}
+
+txSubmissionServer ::
+    MVar MockNodeServerChainState
+ -> TxSubmission.LocalTxSubmissionServer (C.Tx C.BabbageEra) String IO ()
+txSubmissionServer state = txSubmissionState
+    where
+      txSubmissionState :: TxSubmission.LocalTxSubmissionServer (C.Tx C.BabbageEra) String IO ()
+      txSubmissionState =
+        TxSubmission.LocalTxSubmissionServer {
+          TxSubmission.recvMsgSubmitTx =
+            \tx -> do
+                modifyMVar_ state (pure . over txPool (addTxToPool (CardanoEmulatorEraTx tx)))
+                return (TxSubmission.SubmitSuccess, txSubmissionState)
+        , TxSubmission.recvMsgDone     = ()
+        }
